Fixed 'hideCurrentMaterialBanner' and refactored 'AudioPlayerBottomSheet' into a separate widget.
