+++ lib/controllers/internet_chacker.dart
@@ -9,32 +9,46 @@ import 'package:flutter/material.dart';
 
 import '../generated/locale_keys.g.dart'; 
 
-class InternetConnectionController extends ChangeNotifier { 
-
+class InternetConnectionController extends ChangeNotifier {
   late final Connectivity _connectivity;
   InternetConnectionController(this._connectivity);
 
-
   StreamSubscription? streamSubscription;
   bool isConnected = true;
+  ValueNotifier<bool> internetNotifier = ValueNotifier<bool>(true);
 
-  void listenTonetworkChacges(BuildContext context) {
-    streamSubscription = _connectivity.onConnectivityChanged.listen((result) {
+  void listenToNetworkChanges(BuildContext context) {
+    // Check the initial network connectivity status
+    _connectivity.checkConnectivity().then((result) {
       if (result == ConnectivityResult.mobile || result == ConnectivityResult.wifi) {
         isConnected = true;
-        ScaffoldMessenger.of(context).hideCurrentMaterialBanner();
-     
       } else {
         isConnected = false;
         CustomBanner.showBanner(
           context,
-         LocaleKeys.noInternetConnection.tr(),
+          LocaleKeys.noInternetConnection.tr(),
           Colors.grey.shade800,
           const Duration(days: 100),
         );
-     
       }
-      // Instead of `update()`, use `notifyListeners()` to notify listeners of the change.
+      internetNotifier.value = isConnected;
+      notifyListeners();
+    });
+
+    streamSubscription = _connectivity.onConnectivityChanged.listen((result) {
+      if (result == ConnectivityResult.mobile || result == ConnectivityResult.wifi) {
+        isConnected = true;
+        ScaffoldMessenger.of(context).hideCurrentMaterialBanner();
+      } else {
+        isConnected = false;
+        CustomBanner.showBanner(
+          context,
+          LocaleKeys.noInternetConnection.tr(),
+          Colors.grey.shade800,
+          const Duration(days: 100),
+        );
+      }
+      internetNotifier.value = isConnected;
       notifyListeners();
     });
   }
