+++ lib/core/custom_banner.dart
@@ -6,22 +6,32 @@ class CustomBanner {
   static showBanner(
       BuildContext context, String message, Color bgColor, Duration? duration) {
     ScaffoldMessenger.of(context).showMaterialBanner(MaterialBanner(
-      actions: [Container()],
+      actions: [
+        IconButton(
+          onPressed: () {
+            ScaffoldMessenger.of(context).hideCurrentMaterialBanner();
+          },
+          icon: const Icon(
+            Icons.close_outlined,
+            color: Colors.white,
+          ),
+        ),
+      ],
       backgroundColor: bgColor,
-      content:  SizedBox(
+      content: SizedBox(
         width: double.infinity,
         child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
             const Icon(
               Icons.cloud_off_outlined,
               color: Colors.white,
               size: 20,
             ),
-            const SizedBox(
-              width: 16,
+            Text(
+              message,
+              style: const TextStyle(color: Colors.white, fontSize: 16),
             ),
-            Text(message, style: const TextStyle(color: Colors.white, fontSize: 16),)
           ],
         ),
       ),
